% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainclass.R
\name{check_nps}
\alias{check_nps}
\title{Check validity of an \code{nps} object}
\usage{
check_nps(object)
}
\arguments{
\item{object}{A \code{nps} instance}
}
\description{
Check validity of an \code{nps} object
}

