% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters-getters.R
\name{getters}
\alias{bottom}
\alias{getters}
\alias{top}
\title{Getters}
\usage{
top(object)

bottom(object)
}
\arguments{
\item{object}{An \code{nps} object}
}
\description{
Getters

Retrieve the values corresponding to the "Promoters" category

Retrieve the values corresponding to the "Promoters" category
}

