% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainclass.R
\docType{class}
\name{nps-class}
\alias{nps-class}
\title{An S4 class to work with NPS data}
\description{
An S4 class to work with NPS data
}
\section{Slots}{

\describe{
\item{\code{values}}{A numeric vector with the answer values}

\item{\code{top}}{A numeric vector with the values in the "Promoters" category}

\item{\code{bottom}}{A numeric vector with the values in the "Detractors" category}
}}

