% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainclass.R
\name{nps}
\alias{nps}
\title{Instantiate an \code{nps} object}
\usage{
nps(x, top = 9:10, bottom = 0:6)
}
\arguments{
\item{x}{A numeric vector with the answer values}

\item{top}{A numeric vector with the values in the "Promoters" category}

\item{bottom}{A numeric vector with the values in the "Detractors" category}
}
\description{
Instantiate an \code{nps} object
}
\examples{
nps(sample(0:10, size=100, replace=TRUE))
}

