% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{methods}
\name{score}
\alias{score}
\alias{score,nps-method}
\title{Calculation of NPS and estimated standard error}
\usage{
score(object, ...)

\S4method{score}{nps}(object, boot = FALSE, R = 999, ...)
}
\arguments{
\item{object}{An object of class \code{nps}}

\item{...}{Additional arguments (currently not used)}

\item{boot}{A logical flag for whether to use bootstrap for the calculation of the standard error. Default is FALSE.}

\item{R}{Number of bootstrap replications}
}
\value{
A vector with the estimated score and the standard error
}
\description{
Calculation of the score and its standard error. The function allows for the calculation of the standard error using both its analytical derivation and through non-parametric bootstrap.
}
\examples{

\dontrun{score(x, boot=TRUE, R=999)}
}

