% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters-getters.R
\name{setters}
\alias{bottom<-}
\alias{setters}
\alias{top<-}
\title{Setters}
\usage{
top(object) <- value

bottom(object) <- value
}
\arguments{
\item{object}{An \code{nps} object}

\item{value}{The new categories}
}
\value{
An \code{nps} object
}
\description{
Setter for the top and bottom categories
}

