% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters-getters.R
\docType{methods}
\name{[,nps-method}
\alias{[,nps-method}
\title{Subset an NPS vector}
\usage{
\S4method{[}{nps}(x, i)
}
\arguments{
\item{x}{An \code{nps} object}

\item{i}{An index for the selection}
}
\description{
Subset an NPS vector
}

