% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,nps-method}
\alias{summary,nps-method}
\title{Summary of NPS values}
\usage{
\S4method{summary}{nps}(object)
}
\arguments{
\item{object}{An object of class \code{nps}}
}
\value{
A frequency table with counts for each category
}
\description{
Summary table of the number of respondents in each category
}

