% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainclass.R, R/nps.R
\docType{package}
\name{nps}
\alias{nps}
\alias{nps-package}
\title{Instantiate an \code{nps} object}
\usage{
nps(x, top = 9:10, bottom = 0:6)
}
\arguments{
\item{x}{A numeric vector with the answer values}

\item{top}{A numeric vector with the values in the "Promoters" category}

\item{bottom}{A numeric vector with the values in the "Detractors" category}
}
\description{
Instantiate an \code{nps} object

nps -- A package for handling NPS data
}
\section{Caution}{

This is a package is not intended to be used in a production environment. It was designed as a tool for the "R Development" class and it explores a number of features like S4 classes, C/C++ integration, or unit testing in addition to some programming tools.
}

\section{Contents}{

The package contains utilities to handle and analyze data coming from a likelihood to recommend question.
}
\examples{
nps(sample(0:10, size=100, replace=TRUE))
}
\author{
Gonzalo Rivero
}

