% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters-getters.R
\name{setters}
\alias{bottom<-}
\alias{setters}
\alias{top<-}
\title{Setters}
\usage{
top(object) <- value

bottom(object) <- value
}
\arguments{
\item{object}{An \code{nps} object}

\item{value}{A vector with the values of the top or bottom category}
}
\description{
Setters

Change the values corresponding to the "Promoters" category

Change the values corresponding to the "Detractors" category
}

